Change netcore31 chatsample to default chatsample and add messagepack dep
