+++ AzureSignalR.sln
@@ -54,19 +54,19 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.Man
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ChatSample", "ChatSample", "{C965ED06-6A17-4329-B3C6-811830F4F4ED}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample", "samples\ChatSample\ChatSample\ChatSample.csproj", "{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample.CSharpClient", "samples\ChatSample\ChatSample.CSharpClient\ChatSample.CSharpClient.csproj", "{C51AA61E-5210-4545-B49E-61255C136B83}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.Tests.Common", "test\Microsoft.Azure.SignalR.Tests.Common\Microsoft.Azure.SignalR.Tests.Common.csproj", "{981419D7-5199-4829-ABE2-421D05EF7D8C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample.CoreApp3", "samples\ChatSample\ChatSample.CoreApp3\ChatSample.CoreApp3.csproj", "{1F6D9B7A-36A0-4B71-825B-77F832B425A5}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.E2ETests", "test\Microsoft.Azure.SignalR.E2ETests\Microsoft.Azure.SignalR.E2ETests.csproj", "{B8BA8F89-8028-4691-82C6-8F9E946B6B22}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.SignalR.AspNet.E2ETests", "test\Microsoft.Azure.SignalR.AspNet.E2ETests\Microsoft.Azure.SignalR.AspNet.E2ETests.csproj", "{5D97A2CB-DCD6-4865-9C39-34687A14AAE1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorAppSample", "samples\BlazorAppSample\BlazorAppSample.csproj", "{CDE3C0B5-CD0B-42DE-9CE1-638266A2A1AE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample", "samples\ChatSample\ChatSample\ChatSample.csproj", "{63BB9836-126B-4E65-92B9-019438185E41}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChatSample.NetCore21", "samples\ChatSample\ChatSample.NetCore21\ChatSample.NetCore21.csproj", "{1D06C79C-4E97-40F6-A860-092941922EFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -126,10 +126,6 @@ Global
 		{493CE126-D147-434E-ACF8-675E89D0A981}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{493CE126-D147-434E-ACF8-675E89D0A981}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{493CE126-D147-434E-ACF8-675E89D0A981}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C51AA61E-5210-4545-B49E-61255C136B83}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C51AA61E-5210-4545-B49E-61255C136B83}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C51AA61E-5210-4545-B49E-61255C136B83}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -138,10 +134,6 @@ Global
 		{981419D7-5199-4829-ABE2-421D05EF7D8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{981419D7-5199-4829-ABE2-421D05EF7D8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{981419D7-5199-4829-ABE2-421D05EF7D8C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1F6D9B7A-36A0-4B71-825B-77F832B425A5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F6D9B7A-36A0-4B71-825B-77F832B425A5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F6D9B7A-36A0-4B71-825B-77F832B425A5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F6D9B7A-36A0-4B71-825B-77F832B425A5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B8BA8F89-8028-4691-82C6-8F9E946B6B22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B8BA8F89-8028-4691-82C6-8F9E946B6B22}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8BA8F89-8028-4691-82C6-8F9E946B6B22}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -154,6 +146,14 @@ Global
 		{CDE3C0B5-CD0B-42DE-9CE1-638266A2A1AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CDE3C0B5-CD0B-42DE-9CE1-638266A2A1AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDE3C0B5-CD0B-42DE-9CE1-638266A2A1AE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{63BB9836-126B-4E65-92B9-019438185E41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63BB9836-126B-4E65-92B9-019438185E41}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63BB9836-126B-4E65-92B9-019438185E41}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63BB9836-126B-4E65-92B9-019438185E41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D06C79C-4E97-40F6-A860-092941922EFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D06C79C-4E97-40F6-A860-092941922EFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D06C79C-4E97-40F6-A860-092941922EFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D06C79C-4E97-40F6-A860-092941922EFA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -174,13 +174,13 @@ Global
 		{8DC103BA-7404-479F-BF18-90EF3EC3CEA9} = {DA69F624-5398-4884-87E4-B816698CDE65}
 		{493CE126-D147-434E-ACF8-675E89D0A981} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
 		{C965ED06-6A17-4329-B3C6-811830F4F4ED} = {C4BC9889-B49F-41B6-806B-F84941B2549B}
-		{3F9B3FD8-A54F-4AB4-B9D4-D4D55686A0E9} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
 		{C51AA61E-5210-4545-B49E-61255C136B83} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
 		{981419D7-5199-4829-ABE2-421D05EF7D8C} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
-		{1F6D9B7A-36A0-4B71-825B-77F832B425A5} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
 		{B8BA8F89-8028-4691-82C6-8F9E946B6B22} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
 		{5D97A2CB-DCD6-4865-9C39-34687A14AAE1} = {2429FBD8-1FCE-4C42-AA28-DF32F7249E77}
 		{CDE3C0B5-CD0B-42DE-9CE1-638266A2A1AE} = {C4BC9889-B49F-41B6-806B-F84941B2549B}
+		{63BB9836-126B-4E65-92B9-019438185E41} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
+		{1D06C79C-4E97-40F6-A860-092941922EFA} = {C965ED06-6A17-4329-B3C6-811830F4F4ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7945A4E4-ACDB-4F6E-95CA-6AC6E7C2CD59}
