--- samples/ChatSample/ChatSample.CoreApp3/Startup.cs
@@ -1,36 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace ChatSample.CoreApp3
-{
-    public class Startup
-    {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddMvc();
-            services.AddSignalR();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            app.UseFileServer();
-            app.UseRouting();
-            app.UseAuthorization();
-
-            app.UseEndpoints(routes =>
-            {
-                routes.MapHub<Chat>("/chat");
-                routes.MapHub<NotificationHub>("/notificatons");
-            });
-        }
-    }
-}
