--- samples/ChatSample/ChatSample.CoreApp3/appsettings.json
@@ -1,15 +1,0 @@
-{
-  "Azure": {
-    "SignalR": {
-      "Enabled": true,
-      "ConnectionString": ""
-    }
-  },
-  "Logging": {
-    "LogLevel": {
-      "Default": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*"
-}
