+++ samples/ChatSample/ChatSample.NetCore21/Dockerfile
@@ -6,7 +6,7 @@ FROM mcr.microsoft.com/dotnet/core/sdk:3.0
 
 COPY ./ /home/signalr-src
 
-RUN cd /home/signalr-src/samples/ChatSample/ChatSample.CoreApp3 && \
+RUN cd /home/signalr-src/samples/ChatSample/ChatSample && \
     dotnet build && \
     dotnet publish -r ubuntu.16.04-x64 -c Release -o /home/build/
 
@@ -14,7 +14,7 @@ RUN cd /home/signalr-src/samples/ChatSample/ChatSample.CoreApp3 && \
 # Final stage
 ##################################################
 
-FROM mcr.microsoft.com/dotnet/core/runtime:3.0
+FROM mcr.microsoft.com/dotnet/core/runtime:2.1
 
 COPY --from=0 /home/build/ /home/SignalR
 
@@ -22,4 +22,4 @@ WORKDIR /home/SignalR
 
 EXPOSE 5050
 
-CMD ["./ChatSample.CoreApp3"]
+CMD ["./ChatSample"]
