+++ samples/ChatSample/ChatSample.NetCore21/Hub/BenchHub.cs
