+++ samples/ChatSample/ChatSample.NetCore21/Hub/Chat.cs
