+++ samples/ChatSample/ChatSample.NetCore21/LoggerExtension.cs
