+++ samples/ChatSample/ChatSample.NetCore21/appsettings.json
@@ -1,0 +1,13 @@
+﻿{
+  "Azure": {
+    "SignalR": {
+      "ConnectionString": ""
+    }
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.Azure.SignalR": "Information"
+    }
+  }
+}
