+++ samples/ChatSample/ChatSample.NetCore21/package.json
