+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/css/site.css
@@ -31,6 +31,14 @@ textarea:focus {
     background: #87CEFA;
 }
 
+.error {
+    color: red
+}
+
+.success {
+    color: green
+}
+
 .broadcast-message {
     display: inline-block;
     background: yellow;
