+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/favicon.ico
