+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/scripts/bootstrap.js
