+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/scripts/jquery-1.10.2.js
