+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/scripts/msgpack5.js
