+++ samples/ChatSample/ChatSample.NetCore21/wwwroot/scripts/signalr.js
