+++ samples/ChatSample/ChatSample/Chat.cs
