+++ samples/ChatSample/ChatSample/appsettings.Development.json
