+++ samples/ChatSample/ChatSample/appsettings.json
@@ -1,13 +1,15 @@
-﻿{
+{
   "Azure": {
     "SignalR": {
+      "Enabled": true,
       "ConnectionString": ""
     }
   },
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
-      "Microsoft.Azure.SignalR": "Information"
+      "Default": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
     }
-  }
+  },
+  "AllowedHosts": "*"
 }
